Update README.md with current info
The README.md file pointed to the old PPA archive, which is deprecated.
The recommended package was also changed from maratona-meta to
maratona-desktop. Some changes concerning Markdown formatting were
performed to improve the file readability. The older command line
suggestion to use the apt-get was replaced by the newer apt.

- replace old PPA link
- replace maratona-meta with maratona-desktop
- add extra formatting in the file
- replace apt-get by apt